Fix a bug with parsing strings to enums when getting capabilities during login.
When parsing the response from CAPABILITIES we match the list returned
from the server against a list of known capabilties which we store in an
enum.

The enum has user defined attribute strings to match the server strings
and we iteratively match them using a static foreach and
`__traits(getAttributes)`.

A recent compiler fix changed the way `__traits(getAttributes)` works
with enums which broke our code.